Fix type of id.

+++ api/src/main/kotlin/com/okeicalm/simpleJournalEntry/entity/Account.kt
@@ -1,10 +1,9 @@
 package com.okeicalm.simpleJournalEntry.entity
 
-import com.expediagroup.graphql.generator.scalars.ID
 import com.okeicalm.simpleJournalEntry.tables.pojos.Accounts
 
 data class Account(
-    val id: Int = 0,
+    val id: Long = 0,
     val code: String,
     val name: String,
     val elementType: Int,
